Extract DurationFormatter
